+++ calypso_anemometer/telemetry/common.py
@@ -32,11 +32,13 @@ class NetworkTelemetry:
             self.socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEPORT, 1)
             self.socket.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
 
-    def send(self, payload: str):
+    def send(self, payload: str, newline=True):
         logger.info(f"Sending message to {self.protocol.value}://{self.host}:{self.port}. {payload}")
         address = (self.host, self.port)
         if isinstance(payload, str):
             payload = payload.encode("utf-8")
+        if newline:
+            payload += b"\n"
         if self.protocol == NetworkProtocol.TCP:
             self.socket.connect(address)
             self.socket.send(payload)